wip: converting timeouttype to enum

+++ common/persistence/sql/sqlExecutionManagerUtil.go
@@ -894,12 +894,12 @@ func createTimerTasks(
 			switch t := task.(type) {
 			case *p.DecisionTimeoutTask:
 				info.EventId = t.EventID
-				info.TimeoutType = int32(t.TimeoutType)
+				info.TimeoutType = t.TimeoutType
 				info.ScheduleAttempt = t.ScheduleAttempt
 
 			case *p.ActivityTimeoutTask:
 				info.EventId = t.EventID
-				info.TimeoutType = int32(t.TimeoutType)
+				info.TimeoutType = t.TimeoutType
 				info.ScheduleAttempt = t.Attempt
 
 			case *p.UserTimerTask:
@@ -911,7 +911,7 @@ func createTimerTasks(
 
 			case *p.WorkflowBackoffTimerTask:
 				info.EventId = t.EventID
-				info.TimeoutType = int32(t.TimeoutType)
+				info.TimeoutType = t.TimeoutType
 
 			case *p.WorkflowTimeoutTask:
 				// noop
